synq: use uint32_t for token types/lengths in C API + fix clippy warnings
- Change all public C headers to use uint32_t (token types, lengths,
  counts, capacities, boolean flags) and int32_t for PARSE_* return
  codes that can be -1
- grammar.h get_token fn ptr keeps int* (matches extracted SQLite C);
  token_wrapped.c casts uint32_t* → int* at the boundary
- Update buildtools codegen templates (c_dialect.rs, parser_pipeline.rs)
  so regenerated sqlite_parse.h/c use uint32_t throughout
- Fix remaining Rust clippy warnings: unused import, needless lifetime,
  unwrap → expect, cast_possible_truncation allow, iter→to_vec,
  missing SAFETY comments, semicolon_if_nothing_returned

+++ syntaqlite-syntax/csrc/tokenizer.c
@@ -35,18 +35,18 @@ void syntaqlite_tokenizer_reset(SyntaqliteTokenizer* tok,
   tok->offset = 0;
 }
 
-int syntaqlite_tokenizer_next(SyntaqliteTokenizer* tok, SyntaqliteToken* out) {
+uint32_t syntaqlite_tokenizer_next(SyntaqliteTokenizer* tok, SyntaqliteToken* out) {
   if (tok->offset >= tok->len) {
     return 0;
   }
 
-  int token_type = 0;
+  uint32_t token_type = 0;
   int64_t token_len = SynqSqliteGetTokenVersionWrapped(
       &tok->env, (const unsigned char*)tok->source + tok->offset, &token_type);
 
   out->text = tok->source + tok->offset;
   out->length = (uint32_t)token_len;
-  out->type = (uint32_t)token_type;
+  out->type = token_type;
 
   tok->offset += (uint32_t)token_len;
   return 1;
